moved status endpoint to index controller
