+++ sqlx_user_crud/src/controller/index_controller.rs
@@ -1,0 +1,17 @@
+use actix_web::{get, web, Responder, HttpResponse};
+use super::AppState;
+use super::log_request;
+use std::sync::Mutex;
+
+pub fn init(cfg: &mut web::ServiceConfig) {
+    cfg.service(status);
+}
+
+// TODO: move to index controller_test
+#[get("/status")]
+async fn status(data: web::Data<AppState<'_>>) -> impl Responder {
+    log_request("GET: /status", &data.connections);
+
+    HttpResponse::Ok()
+        .body("I am up")
+}
